refactor: 拆分stt转换模块hooks组件和配置文件
