+++ src/renderer/components/states/agentVoiceInput/components/AgentVoiceInputView.tsx
@@ -1,0 +1,47 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+import type { ReactElement } from 'react';
+
+interface AgentVoiceInputViewProps {
+  statusText: string;
+  transcript: string;
+  textRef: React.RefObject<HTMLDivElement | null>;
+}
+
+export function AgentVoiceInputView(props: AgentVoiceInputViewProps): ReactElement {
+  const { statusText, transcript, textRef } = props;
+
+  return (
+    <div className="agent-voice-input-content">
+      <div className="agent-voice-input-status">
+        <div className="agent-voice-input-indicator">
+          <span className="agent-voice-input-dot" />
+          <span className="agent-voice-input-dot" />
+          <span className="agent-voice-input-dot" />
+        </div>
+        <span className="agent-voice-input-label">{statusText}</span>
+      </div>
+      <div className="agent-voice-input-text" ref={textRef}>
+        <span className="agent-voice-input-transcript">{transcript || '...'}</span>
+      </div>
+    </div>
+  );
+}
