+++ src/renderer/components/states/agentVoiceInput/config/agentVoiceInputConfig.ts
@@ -1,0 +1,32 @@
+/*
+ * eIsland - A sleek, Apple Dynamic Island inspired floating widget for Windows, built with Electron.
+ * https://github.com/JNTMTMTM/eIsland
+ *
+ * Copyright (C) 2026 JNTMTMTM
+ * Copyright (C) 2026 pyisland.com
+ *
+ * Original author: JNTMTMTM[](https://github.com/JNTMTMTM)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+export const AGENT_VOICE_INITIAL_STATUS = '正在连接ASR服务';
+export const AGENT_VOICE_STT_LANGUAGE = 'zh-CN';
+export const AGENT_VOICE_FRAME_SIZE = 320;
+export const AGENT_VOICE_MAX_RECORDING_MS = 60_000;
+
+export const AGENT_VOICE_AUDIO_CONSTRAINTS: MediaTrackConstraints = {
+  channelCount: 1,
+  sampleRate: 16000,
+  echoCancellation: true,
+  noiseSuppression: true,
+  autoGainControl: true,
+};
